requirements.txt
environs==9.5.0
python-telegram-bot==20.4
django==4.2.3